Port bugfixes from LidarDEV, refactor more to com.spartronics4915.lidar
