+++ src/main/java/com/spartronics4915/lib/lidar/LidarPoint.java
@@ -40,8 +40,7 @@ class LidarPoint
                                               Math.sin(radians) * this.distance);
         if(robotPose != null)
         {
-            robotPose.transformBy(Pose2d.fromTranslation(x2d));
-            return robotPose.getTranslation();
+            return robotPose.transformBy(Pose2d.fromTranslation(x2d)).getTranslation();
         }
         else
             return x2d;
